Split test_benchmarks.py -> test_benchmarks.py, test_runner.py
